+++ public/index.html
@@ -17,7 +17,6 @@
 	<script src="js/login.js" defer></script>
 	<script src="js/index.js" defer></script>
 	<script src="js/chat.js" defer></script>
-	<script src="js/tick.js" defer></script>
 
 
 	<!-- CSS -->
