--- public/js/tick.js
@@ -1,5 +1,0 @@
-'use strict';
-
-socket.on('tick', (msg) => {
-    console.log(msg);
-})
